+++ packages/typegen/Users/ericluce/Documents/Code/work/proofkit/packages/typegen/tests/typegen-output/config1/client/index.ts
@@ -1,0 +1,1 @@
+export { client as testLayoutLayout } from "./testLayout";